+++ lib/basic/widget.dart
@@ -1,8 +1,7 @@
 import 'package:objd/build/build.dart';
 
 /// A widget is the base element for basically everything in objD.
-class Widget {
-
-   generate(Context context) {}
+abstract class Widget {
+   generate(Context context);
    toMap() => null;
 }